Modulize course classfier
