+++ routes/util/courseClassfier.js
@@ -1,0 +1,73 @@
+let classfier = function(courses, sid_dept){
+    let calcPoint = (acc, e) => (acc += parseFloat(e[2]));
+    let calcPointOfObj = obj => {
+      obj["p"] = Object.keys(obj).reduce((acc, x) => (acc += obj[x]), 0);
+    };
+    let passed = courses.filter(x => x[3] >= 60);
+
+    //共同必修
+    let common = passed.filter(x => x[0].match(/^GR.+/));
+    let chinese = common.filter(x => x[1].match(/^中文.+/));
+    let english = common.filter(x => x[1].match(/^英語會話與閱讀.+/));
+    let other = common.filter(x => x[1].match(/^(?!中文|英語會話與閱讀)/));
+    common = { chinese, english, other };
+    let common_p = {
+      chinese: chinese.reduce(calcPoint, 0),
+      english: english.reduce(calcPoint, 0),
+      other: other.reduce(calcPoint, 0)
+    };
+    calcPointOfObj(common_p);
+
+    //核心通識
+    let cc = passed.filter(x => x[0].match(/^CC.+/));
+    let cc_p = cc.reduce(calcPoint, 0);
+
+    //博雅通識
+    let li = passed.filter(x => x[0].match(/^LI.+/));
+    let sc = passed.filter(x => x[0].match(/^SC.+/));
+    let so = passed.filter(x => x[0].match(/^SO.+/));
+    let by = { li, sc, so };
+    let by_p = {
+      li: li.reduce(calcPoint, 0),
+      sc: sc.reduce(calcPoint, 0),
+      so: so.reduce(calcPoint, 0)
+    };
+    calcPointOfObj(by_p);
+
+    // 系必修/系選修
+    let dept = passed.filter(x =>
+      x[0].match(new RegExp(`^${sid_dept[0]}.+`))
+    );
+    let dept_compulsory = dept.filter(x => x[4]);
+    let dept_choose = dept.filter(x => !x[4]);
+    dept = { compulsory: dept_compulsory, choose: dept_choose };
+    let dept_p = {
+      compulsory: dept_compulsory.reduce(calcPoint, 0),
+      choose: dept_choose.reduce(calcPoint, 0)
+    };
+
+    //剩餘學分
+    let rest = passed.filter(x =>
+      x[0].match(new RegExp(`^(?!GR|CC|LI|SC|SO|${sid_dept[0]})`))
+    );
+    let rest_p = rest.reduce(calcPoint, 0);
+
+    //===指揮挺組合===
+    let doneCourse = { common, cc, by, dept, rest };
+    let donePoint = {
+      common: common_p,
+      cc: cc_p,
+      by: by_p,
+      dept: dept_p,
+      rest: rest_p
+    };
+    donePoint["p"] = passed.reduce(calcPoint, 0);
+
+    //未送成績和棄選的
+    let uncommitCourse = courses.filter(x => x[3] === "未送");
+    let failedCourse = courses.filter(x => x[3] < 60);
+
+    return {doneCourse, donePoint, uncommitCourse, failedCourse};
+}
+
+module.exports = classfier;
